#6 #7 #8
added custom view holders for Drive and Event.
added new layouts for the drive and event recycler views.
updated the recycler view function to suite the new holders and layout files.
added some logic to the recycler view function so that there is an option to do something after one of the rows are pressed in the recycler view.

Signed-off-by: daniginsberg <dani.ginsberg@gmail.com>