#17 #18 #19
added the user class, so we will be able to keep the users phone number- currently still havent been able to get the phone number out of the getphoneNumber method.
added the needed add and get for the user, as mentioned before the get method is not yet behaving as we want it to. there is a TODO about it.
added firebase storage to the project and created a function in order to add pictures to the storage database.

Signed-off-by: daniginsberg <dani.ginsberg@gmail.com>