#18 #25 #26
added a method so that the user phone number can be extracted and showed in a given text view.
added a get method for images from the firebase storage platform.
changed the get and store image methods so that they support storing event photos, and facebook profile pictures
in the firebase storage platform.

Signed-off-by: daniginsberg <dani.ginsberg@gmail.com>